Daebak

+++ src/Wiskunde.java
@@ -1,6 +1,6 @@
 public class Wiskunde {
-    public int max(int a, int ... rest) {
-        int grootste=a;
+    public int max(byte a, int ... rest) {
+        byte grootste=a;
         return grootste;
     }
 }
